Slight Refactor to separate the color handling from ConnectedWingAlike
