+++ common/src/main/java/org/valkyrienskies/clockwork/ClockWorkItems.java
@@ -2,23 +2,16 @@ package org.valkyrienskies.clockwork;
 
 import com.simibubi.create.AllTags;
 import com.simibubi.create.content.AllSections;
-import com.simibubi.create.content.curiosities.CombustibleItem;
 import com.simibubi.create.foundation.data.AssetLookup;
-import com.simibubi.create.foundation.data.CreateRegistrate;
 import com.tterrag.registrate.util.entry.ItemEntry;
-import net.minecraft.world.item.BucketItem;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.Rarity;
 import org.valkyrienskies.clockwork.content.curiosities.tools.bluperglue.BluperGlueItem;
 import org.valkyrienskies.clockwork.content.curiosities.tools.gravitron.GravitronItem;
 import org.valkyrienskies.clockwork.content.curiosities.tools.gravitron.GravitronItemRenderer;
-import org.valkyrienskies.clockwork.content.curiosities.tools.pastrymaker.PastrymakerItem;
-import org.valkyrienskies.clockwork.content.curiosities.tools.pastrymaker.PastrymakerItemRenderer;
 import org.valkyrienskies.clockwork.content.curiosities.tools.welder.WelderItem;
 import org.valkyrienskies.clockwork.content.curiosities.tools.welder.WelderItemRenderer;
-import org.valkyrienskies.clockwork.content.materials.solids.soda.SodaBottleItem;
 import org.valkyrienskies.clockwork.content.materials.solids.stratodonut.StratodonutItem;
-import org.valkyrienskies.clockwork.content.physicalities.motion.wing.WingBlockItem;
 import org.valkyrienskies.clockwork.util.builder.ClockworkRegistrate;
 
 import static com.simibubi.create.content.AllSections.MATERIALS;
