+++ common/src/main/java/org/valkyrienskies/clockwork/content/physicalities/motion/wing/DyedWingBlockItem.java
@@ -6,27 +6,23 @@ import net.minecraft.network.chat.Component;
 import net.minecraft.network.chat.MutableComponent;
 import net.minecraft.network.chat.Style;
 import net.minecraft.network.chat.TextComponent;
-import net.minecraft.world.InteractionResult;
 import net.minecraft.world.item.BlockItem;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.context.BlockPlaceContext;
-import net.minecraft.world.item.context.UseOnContext;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.LayeredCauldronBlock;
 import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
 import org.jetbrains.annotations.Nullable;
 import org.valkyrienskies.clockwork.ClockWorkBlockEntities;
 import org.valkyrienskies.clockwork.content.materials.solids.colorblock.ColorBlockEntity;
-import org.valkyrienskies.clockwork.util.blocktype.ConnectedWingAlike;
 
 import java.util.List;
 import java.util.Set;
 
-public class WingBlockItem extends BlockItem {
-    public WingBlockItem(Block block, Properties properties) {
+public class DyedWingBlockItem extends BlockItem {
+    public DyedWingBlockItem(Block block, Properties properties) {
         super(block, properties);
     }
